refactor(hindsight): put flow recovery behind a DecodeStrategy trait
The decoder now holds an ordered list of decode strategies and asks
each in turn per matched transaction; the first that returns a flow
wins, so a later strategy is the fallback when an earlier one cannot
decode (e.g. a future calldata strategy behind transfer netting).

The whole current flow recovery becomes the first implementation
(strategies/netting.rs); the sender/maker/venue classification is now
its private TraderShape. What was strategy.rs shrinks to matching.rs:
the receipt-only "is this a solver trade" filter plus vetoes.

Module docs spell out the placement rules: a strategy is a method for
extracting swaps, never a venue or a solver; a venue's quirks all live
in its venues/ module (transfer-based and calldata-based layers);
per-solver calldata formats are match arms inside the method that
reads them.
